Focus on React (#2)
This commit changes `accept-charset` to `acceptCharset` and `auto-complete` to
`autoComplete`. We also have the submit helper props expect to use a submit
button instead of an input, because React doesn't have a `defaultValue` prop
for `<input type="submit">`.

Generally, i'm deciding to target exclusively on React for now. I think we can
fix these for HTML someday.