Make attribute setting explicit. (#4)
With prop_template removing implicit conversations using camelize(:lower), this
commit makes the key formatting the responsibility of form_props. To do this,
we keep a mapping of html attributes => React attributes as provided by React,
and introduce our own with certain options that are needed for frontend components
like `include_hidden`.

The mapping is used right before we pass the keys to prop_template.