Pass `form_props` options to `Inputs::RadioButton`
In order to access inherited options like `controlled`, we need to
pass down the options given to `form_props` to the
`Inputs::RadioButton` we create inside of the `radio_button` helper.

This tweaks that helper to call `objectify_options` similar to how
the other helpers do. The `test_form_with_using_controlled_option`
test was then modified to exercise this helper in order to confirm the
change.